fix: Field Number set value with enter.
